pyproject.toml
[project]
name = "rag-sys"
version = "0.1.0"
description = "Production-grade RAG system with FastAPI and Weaviate"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34,<1.0",
    "weaviate-client>=4.9,<5.0",
    "openai>=1.60,<2.0",
    "sentence-transformers>=3.4,<4.0",
    "langchain-text-splitters>=0.3,<1.0",
    "nltk>=3.9,<4.0",
    "python-multipart>=0.0.18",
    "pypdf>=5.0,<6.0",
    "python-docx>=1.1,<2.0",
    "pydantic-settings>=2.7,<3.0",
    "structlog>=24.4,<26.0",
    "numpy>=1.26,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.25,<1.0",
    "httpx>=0.28,<1.0",
    "ruff>=0.9,<1.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
